#13 refactored max to accept an arbitrary number of input int's
