+++ util/max.go
@@ -1,16 +1,15 @@
-// Package maximumtriplet provides a function to find the highest integer in a list of three.
-package maximumtriplet
+package util
 
-// MaximumTriplet finds the maximum value of three signed integers.
+// Max finds the maximum value of an arbitrary number of signed integers.
 // It returns the maximum signed integer found.
-// Time complexity is O(1).
-func MaximumTriplet(a, b, c *int) int {
-	switch {
-	case *a > *b && *a > *c:
-		return *a
-	case *b > *c:
-		return *b
-	default:
-		return *c
+// Time complexity is O(N).
+func Max(a int, b ...int) int {
+	max := a
+	for _, i := range b {
+		if i > max {
+			max = i
+		}
 	}
+
+	return max
 }
